+++ src/HassModel/NetDaemon.HassModel.CodeGenerator/Extensions/TypeExtensions.cs
@@ -1,62 +1,61 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace NetDaemon.HassModel.CodeGenerator.Extensions
+namespace NetDaemon.HassModel.CodeGenerator.Extensions;
+
+internal static class TypeExtensions
 {
-    internal static class TypeExtensions
+    private static readonly Dictionary<Type, string> _typeToFriendlyName = new()
     {
-        private static readonly Dictionary<Type, string> _typeToFriendlyName = new()
+        { typeof(int), "int" },
+        { typeof(uint), "uint" },
+        { typeof(long), "long" },
+        { typeof(ulong), "ulong" },
+        { typeof(float), "float" },
+        { typeof(double), "double" },
+        { typeof(object), "object" },
+        { typeof(bool), "bool" },
+        { typeof(string), "string" },
+        { typeof(int?), "int?" },
+        { typeof(uint?), "uint?" },
+        { typeof(long?), "long?" },
+        { typeof(ulong?), "ulong?" },
+        { typeof(float?), "float?" },
+        { typeof(double?), "double?" },
+        { typeof(bool?), "bool?" },
+        { typeof(void), "void" }
+    };
+
+    public static string GetFriendlyName(this Type type)
+    {
+        if (_typeToFriendlyName.TryGetValue(type, out var friendlyName))
         {
-            { typeof(int), "int" },
-            { typeof(uint), "uint" },
-            { typeof(long), "long" },
-            { typeof(ulong), "ulong" },
-            { typeof(float), "float" },
-            { typeof(double), "double" },
-            { typeof(object), "object" },
-            { typeof(bool), "bool" },
-            { typeof(string), "string" },
-            { typeof(int?), "int?" },
-            { typeof(uint?), "uint?" },
-            { typeof(long?), "long?" },
-            { typeof(ulong?), "ulong?" },
-            { typeof(float?), "float?" },
-            { typeof(double?), "double?" },
-            { typeof(bool?), "bool?" },
-            { typeof(void), "void" }
-        };
-
-        public static string GetFriendlyName(this Type type)
-        {
-            if (_typeToFriendlyName.TryGetValue(type, out var friendlyName))
-            {
-                return friendlyName;
-            }
-
-            friendlyName = type.Name;
-            if (type.IsGenericType)
-            {
-                var backtick = friendlyName.IndexOf('`', StringComparison.InvariantCultureIgnoreCase);
-                if (backtick > 0)
-                {
-                    friendlyName = friendlyName.Remove(backtick);
-                }
-                friendlyName += "<";
-                Type[] typeParameters = type.GetGenericArguments();
-                for (var i = 0; i < typeParameters.Length; i++)
-                {
-                    string typeParamName = typeParameters[i].GetFriendlyName();
-                    friendlyName += (i == 0 ? typeParamName : ", " + typeParamName);
-                }
-                friendlyName += ">";
-            }
-
-            if (type.IsArray)
-            {
-                return type.GetElementType()?.GetFriendlyName() + "[]";
-            }
-
             return friendlyName;
         }
+
+        friendlyName = type.Name;
+        if (type.IsGenericType)
+        {
+            var backtick = friendlyName.IndexOf('`', StringComparison.InvariantCultureIgnoreCase);
+            if (backtick > 0)
+            {
+                friendlyName = friendlyName.Remove(backtick);
+            }
+            friendlyName += "<";
+            Type[] typeParameters = type.GetGenericArguments();
+            for (var i = 0; i < typeParameters.Length; i++)
+            {
+                string typeParamName = typeParameters[i].GetFriendlyName();
+                friendlyName += (i == 0 ? typeParamName : ", " + typeParamName);
+            }
+            friendlyName += ">";
+        }
+
+        if (type.IsArray)
+        {
+            return type.GetElementType()?.GetFriendlyName() + "[]";
+        }
+
+        return friendlyName;
     }
 }
